Decouple shortcuts and apply theme icons
