+++ Passepartout/App/SceneDelegate.swift
@@ -39,59 +39,3 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         handleShortcutItem(shortcutItem)
     }
 }
-
-private extension SceneDelegate {
-    enum ShortcutType: String {
-        case enableVPN
-
-        case disableVPN
-
-        case reconnectVPN
-    }
-
-    func rebuildShortcutItems() {
-        let items: [UIApplicationShortcutItem]
-        if VPNManager.shared.currentState.isEnabled {
-            let toggleItem = UIApplicationShortcutItem(
-                type: ShortcutType.disableVPN.rawValue,
-                localizedTitle: L10n.Shortcuts.Add.Items.DisableVpn.caption
-            )
-            let reconnectItem = UIApplicationShortcutItem(
-                type: ShortcutType.reconnectVPN.rawValue,
-                localizedTitle: L10n.Global.Strings.reconnect
-            )
-            items = [toggleItem, reconnectItem]
-        } else if ProfileManager.shared.hasActiveProfile {
-            let toggleItem = UIApplicationShortcutItem(
-                type: ShortcutType.enableVPN.rawValue,
-                localizedTitle: L10n.Shortcuts.Add.Items.EnableVpn.caption
-            )
-            items = [toggleItem]
-        } else {
-            items = []
-        }
-        UIApplication.shared.shortcutItems = items
-    }
-
-    func handleShortcutItem(_ shortcutItem: UIApplicationShortcutItem) {
-        switch shortcutItem.type {
-        case ShortcutType.enableVPN.rawValue:
-            Task {
-                try await VPNManager.shared.connectWithActiveProfile(toServer: nil)
-            }
-
-        case ShortcutType.disableVPN.rawValue:
-            Task {
-                await VPNManager.shared.disable()
-            }
-
-        case ShortcutType.reconnectVPN.rawValue:
-            Task {
-                await VPNManager.shared.reconnect()
-            }
-
-        default:
-            break
-        }
-    }
-}
